--- common/src/main/java/com/ryazancev/common/dto/notification/NotificationType.java
@@ -1,10 +1,0 @@
-package com.ryazancev.common.dto.notification;
-
-/**
- * @author Oleg Ryazancev
- */
-
-public enum NotificationType {
-
-
-}
