+++ common/src/main/java/com/ryazancev/common/dto/notification/enums/NotificationScope.java
@@ -1,4 +1,4 @@
-package com.ryazancev.common.dto.notification;
+package com.ryazancev.common.dto.notification.enums;
 
 /**
  * @author Oleg Ryazancev
