+++ common/src/main/java/com/ryazancev/common/dto/notification/enums/NotificationStatus.java
@@ -1,4 +1,4 @@
-package com.ryazancev.notification.model;
+package com.ryazancev.common.dto.notification.enums;
 
 /**
  * @author Oleg Ryazancev
