+++ common/src/main/java/com/ryazancev/common/dto/notification/enums/NotificationType.java
@@ -1,0 +1,17 @@
+package com.ryazancev.common.dto.notification.enums;
+
+/**
+ * @author Oleg Ryazancev
+ */
+
+public enum NotificationType {
+    PRIVATE_PRODUCT_REGISTRATION_ACCEPTED,
+    PRIVATE_PRODUCT_REGISTRATION_REJECTED,
+    PRIVATE_ORGANIZATION_REGISTRATION_ACCEPTED,
+    PRIVATE_ORGANIZATION_REGISTRATION_REJECTED,
+    PRIVATE_ACCOUNT_LOCKED,
+    PRIVATE_ACCOUNT_UNLOCKED,
+    PRIVATE_REVIEW_RECEIVED,
+    PRIVATE_PURCHASE_PROCESSED,
+    PUBLIC_NEW_PRODUCT_CREATED
+}
